requirements.txt
fitz
sentence-transformers
pinecone-client
flask

